Arumando comits

+++ ClinicaOdont/src/main/java/com/sistemaClinica/dentista/service/DentistaService.java
@@ -4,6 +4,7 @@ import com.sistemaClinica.dentista.dto.DentistaDTO;
 import com.sistemaClinica.dentista.mapper.DentistaMapper;
 import com.sistemaClinica.dentista.model.Dentista;
 import com.sistemaClinica.dentista.repository.DentistaRepository;
+import com.sistemaClinica.shared.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -25,22 +26,21 @@ public class DentistaService {
                 .collect(Collectors.toList());
     }
 
-    public DentistaDTO buscarPorId(String id) {
+    public DentistaDTO buscarPorId(Integer id) {
         return dentistaRepository.findById(id)
                 .map(dentistaMapper::toDto)
-                .orElse(null);
+                .orElseThrow(() -> new ResourceNotFoundException("Dentista nao encontrado"));
     }
 
     public DentistaDTO salvar(DentistaDTO dentistaDTO) {
         Dentista dentista = dentistaMapper.toEntity(dentistaDTO);
-        // Lógica para gerar ID se for uma nova entidade, se necessário
-        if (dentista.getIdDentista() == null || dentista.getIdDentista().isEmpty()) {
-            // O @GeneratedValue cuidará disso se configurado na entidade
-        }
         return dentistaMapper.toDto(dentistaRepository.save(dentista));
     }
 
-    public void deletar(String id) {
+    public void deletar(Integer id) {
+        if (!dentistaRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Dentista nao encontrado");
+        }
         dentistaRepository.deleteById(id);
     }
 }
